made requested changes

+++ tvtk/pyface/picker.py
@@ -21,7 +21,7 @@ probe for the data at that point.
 # Copyright (c) 2004-2016, Enthought, Inc.
 # License: BSD Style.
 
-from traits.api import HasTraits, Trait, Long, Array, Any, Float, \
+from traits.api import HasTraits, Trait, Int, Array, Any, Float, \
                                 Instance, Range, true, Str, false
 from traitsui.api import View, Group, Item, Handler
 from tvtk.api import tvtk
@@ -60,9 +60,9 @@ class PickedData(HasTraits):
     valid = Trait(false_bool_trait,
                   desc='specifies the validity of the pick event')
     # Id of picked point (-1 implies none was picked)
-    point_id = Long(-1, desc='the picked point ID')
+    point_id = Int(-1, desc='the picked point ID')
     # Id of picked cell (-1 implies none was picked)
-    cell_id = Long(-1, desc='the picked cell ID')
+    cell_id = Int(-1, desc='the picked cell ID')
     # World pick -- this has no ID.
     world_pick = Trait(false_bool_trait,
                        desc='specifies if the pick is a world pick.')
@@ -106,7 +106,7 @@ class PickHandler(HasTraits):
 class DefaultPickHandler(PickHandler):
     """The default handler for the picked data."""
     # Traits.
-    ID = Trait(None, None, Long, desc='the picked ID')
+    ID = Trait(None, None, Int, desc='the picked ID')
 
     coordinate = Trait(None, None, Array('d', (3,)),
                        desc='the coordinate of the picked point')
